requirements.txt
Django==4.2.11
gunicorn==20.1.0
psycopg2==2.9.9
dj_database_url==1.3.0
ipython==8.18.1

# Django Rest Framework
djangorestframework==3.15.1
drf-spectacular==0.27.2
djangorestframework-simplejwt==5.3.1
PyJWT==2.8.0

python-dotenv==0.21.1
pillow==10.3.0

celery==5.4.0
redis
django_celery_beat
ipython==8.18.1